Extract shooting tactics into separate classes
