+++ Model/Gunner.cs
@@ -21,12 +21,13 @@ namespace Vsite.Oom.Battleship.Model
             shipsToShoot = new List<int>(shipLenghts.OrderByDescending(l => l));
             ShootingTactics = ShootingTactics.Random;
             squareTerminator = new SquareTerminator(rows, columns);
+            targetSelect = new RandomShooting(evidenceGrid);
 
         }
         public Square NextTarget() {
             //TODO: implement
 
-            lastTarget = SelectTarget();
+            lastTarget = targetSelect.NextTarget(shipsToShoot[0]);
             return lastTarget;
         }
 
@@ -55,9 +56,6 @@ namespace Vsite.Oom.Battleship.Model
 
             }
             ChangeTactics(hitResult);
-
-            
-          
         }
 
         private void ChangeTactics(HitResult hitResult)
@@ -65,6 +63,7 @@ namespace Vsite.Oom.Battleship.Model
             if(hitResult == HitResult.Sunken)
             {
                 ShootingTactics = ShootingTactics.Random;
+                targetSelect = new RandomShooting(evidenceGrid);
                 return;
             }
             if (hitResult == HitResult.Hit) 
@@ -73,9 +72,11 @@ namespace Vsite.Oom.Battleship.Model
                 {
                     case ShootingTactics.Random:
                         ShootingTactics = ShootingTactics.Surrounding;
+                        targetSelect = new SurroundShooting(evidenceGrid, squaresHit);
                         return;
                     case ShootingTactics.Surrounding:
                         ShootingTactics = ShootingTactics.Inline;
+                        targetSelect = new InlineShooting(evidenceGrid, squaresHit);
                         return;
                     case ShootingTactics.Inline:
                         return;
@@ -83,61 +84,6 @@ namespace Vsite.Oom.Battleship.Model
             }
             
         }
-
-        private Square SelectTarget()
-        {
-            switch (ShootingTactics) {
-                case ShootingTactics.Random:
-                    return SelectRandomly();
-                case ShootingTactics.Surrounding:
-                    return SelectFromAround();
-                case ShootingTactics.Inline:
-                    return SelectInline();
-                default:
-                    Debug.Assert(false);
-                    return null;
-            }
-               
-        }
-        private Square SelectRandomly()
-        {
-            var placements = evidenceGrid.GetAvailablePlacements(shipsToShoot[0]);
-            var allCandidates = placements.SelectMany(seq => seq);
-            var groups = allCandidates.GroupBy(sq => sq);
-            var maxCount = groups.Max(g => g.Count());
-            var largestGroups = groups.Where(g => g.Count() == maxCount);
-            var mostCommon = largestGroups.Select(g => g.Key);
-            if (mostCommon.Count() == 1)
-                return mostCommon.First();
-            int index = random.Next(0, mostCommon.Count());
-            return mostCommon.ElementAt(index);
-        }
-        private Square SelectInline()
-        {
-            var l = evidenceGrid.GetSquaresInLine(squaresHit);
-            if (l.Count() == 1)
-                return l.ElementAt(0).First();
-
-            l.OrderByDescending(ls => ls.Count());
-            int index = random.Next(0, l.Count());
-            return l.ElementAt(index).First();
-
-        }
-        private Square SelectFromAround()
-        {
-            List<IEnumerable<Square>> around = new List<IEnumerable<Square>>();
-            foreach (Direction direction in Enum.GetValues(typeof(Direction))) {
-                var l = evidenceGrid.GetSquaresNextTo(squaresHit.First(), direction);
-                if (l.Count() > 0)
-                    around.Add(l);
-            }
-            if (around.Count == 1)
-                return around[0].First();
-
-            int index = random.Next(0, around.Count);
-            return around[index].First();
-        }
-
 
 
         private Square lastTarget;
@@ -151,6 +97,8 @@ namespace Vsite.Oom.Battleship.Model
         private SortedSquares squaresHit = new SortedSquares();
 
         private ISquareTerminator squareTerminator;
+
+        private ITargetSelect targetSelect;
         public ShootingTactics ShootingTactics { get; private set; }
 
     }
