+++ Model/RandomShooting.cs
@@ -1,0 +1,30 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Vsite.Oom.Battleship.Model
+{
+    public class RandomShooting : ITargetSelect
+    {
+        public RandomShooting(Grid evidenceGrid) {
+            this.evidenceGrid = evidenceGrid;
+        }
+        public Square NextTarget(int shipLenght)
+        {
+            var placements = evidenceGrid.GetAvailablePlacements(shipLenght);
+            var allCandidates = placements.SelectMany(seq => seq);
+            var groups = allCandidates.GroupBy(sq => sq);
+            var maxCount = groups.Max(g => g.Count());
+            var largestGroups = groups.Where(g => g.Count() == maxCount);
+            var mostCommon = largestGroups.Select(g => g.Key);
+            if (mostCommon.Count() == 1)
+                return mostCommon.First();
+            int index = random.Next(0, mostCommon.Count());
+            return mostCommon.ElementAt(index);
+        }
+        private Random random = new Random();
+        private readonly Grid evidenceGrid;
+    }
+}
