+++ Model/SurroundShooting.cs
@@ -1,0 +1,34 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Vsite.Oom.Battleship.Model
+{
+    public class SurroundShooting : ITargetSelect
+    {
+        public SurroundShooting(Grid evidenceGrid,SortedSquares squaresHit) {
+            this.squaresHit = squaresHit;
+            this.evidenceGrid = evidenceGrid;
+        }
+        public Square NextTarget(int shipLenght)
+        {
+            List<IEnumerable<Square>> around = new List<IEnumerable<Square>>();
+            foreach (Direction direction in Enum.GetValues(typeof(Direction)))
+            {
+                var l = evidenceGrid.GetSquaresNextTo(squaresHit.First(), direction);
+                if (l.Count() > 0)
+                    around.Add(l);
+            }
+            if (around.Count == 1)
+                return around[0].First();
+
+            int index = random.Next(0, around.Count);
+            return around[index].First();
+        }
+        private Random random = new Random();
+        private SortedSquares squaresHit;
+        private readonly Grid evidenceGrid;
+    }
+}
